project_4/requirements.txt
pwntools==4.12.0
capstone==5.0.1
filebytes==0.10.2
keystone-engine==0.9.2
ropper==1.13.8
